feat(starter library) detail page design changes (#8568)
* [www/starters] start detail header tweaks

* [www/starters] rearranging and tweaks

* [www/starters] rearrange components to make it more readable

* [www/starters] reorder detail page ui on downsize

* [www/starters] file renaming away from showcase, toward library

* Lint, format

* import React
* remove unused imports

* Remove meta/header/screenshot borders, decrease title/meta vertical whitespace

* Rm one more unused import

* [www/starters] conflict resolution from sourcing merge

* [www/starter] rm old template file
