+++ www/src/assets/logos.js
@@ -5,6 +5,7 @@ import ReactJSIcon from "./react.svg"
 import SegmentIcon from "./segment.svg"
 import FormidableIcon from "./formidable.svg"
 import FabricIcon from "./fabric.svg"
+import NetlifyIcon from "./netlify.svg"
 
 export {
   JSIcon,
@@ -14,4 +15,5 @@ export {
   SegmentIcon,
   FormidableIcon,
   FabricIcon,
+  NetlifyIcon,
 }
